@@ -9,7 +9,9 @@
     "utils"
   ],
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
+    "start": "npm run watch",
+    "watch": "tsc -p tsconfig.json -w",
+    "lint": "tslint --project tsconfig.json -t stylish",
     "prebuild": "rimraf dist && npm run lint",
     "build": "tsc -p tsconfig.json",
     "magic-publish": "git checkout master && git pull && git merge develop && git push && npm run build && npm version minor && npm publish && git push && git checkout develop && git merge master && git push"
@@ -24,18 +26,19 @@
   },
   "devDependencies": {
     "@types/chalk": "^0.4.31",
-    "@types/gulp": "^3.8.32",
-    "@types/node": "0.0.3",
+    "@types/gulp": "^4.0.2",
+    "@types/node": "^7.0.0",
     "require-dir": "^0.3.1",
     "rimraf": "^2.5.4",
     "tslint": "^4.0.0",
-    "yargs": "^6.3.0"
+    "yargs": "^7.0.2"
   },
   "dependencies": {
+    "@speedy/build-tools": "^0.1.0",
     "chalk": "^1.1.3",
     "cross-spawn-promise": "^0.9.0",
-    "rollup": "^0.37.0",
-    "typescript": "^2.1.0"
+    "rollup": "^0.41.6",
+    "typescript": "^2.2.0"
   },
   "peerDependencies": {
     "gulp": "^3.9.1"
