@@ -1,2 +1,3 @@
-export * from "./transpile";
 export * from "./gulp-utils";
+export * from "./lint";
+export * from "./transpile";
